refactor(apply-i18n): widgets
