+++ assets/i18n/kr.i18n.yaml
@@ -85,7 +85,8 @@ name_text_count: "($count / 20)"
 wallet_subtitle: "$name의 $index번 키"
 
 forgot_password: "비밀번호가 기억나지 않나요?"
-qr_link: "네트워크가 활성화된 기기에서 QR 코드를 스캔하시거나 위의 URL 주소로 접속해 주세요."
+scan_qr_url_link: "네트워크가 활성화된 기기에서 QR 코드를 스캔하시거나 위의 URL 주소로 접속해 주세요."
+scan_qr_email_link: "네트워크가 활성화된 기기에서 QR 코드를 스캔하시거나 위의 주소로 메일을 전송해 주세요"
 developer_option: "개발자 옵션"
 
 # lib
