+++ lib/screens/app_unavailable_notification_screen.dart
@@ -4,7 +4,7 @@ import 'package:coconut_vault/localization/strings.g.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:coconut_vault/styles.dart';
-import 'package:coconut_vault/widgets/high-lighted-text.dart';
+import 'package:coconut_vault/widgets/highlighted_text.dart';
 
 class AppUnavailableNotificationScreen extends StatefulWidget {
   const AppUnavailableNotificationScreen({super.key});
