+++ lib/widgets/guide_dialog.dart
@@ -1,3 +1,4 @@
+import 'package:coconut_vault/localization/strings.g.dart';
 import 'package:flutter/material.dart';
 import 'package:coconut_vault/styles.dart';
 import 'package:coconut_vault/widgets/button/action_bottom_button.dart';
@@ -17,7 +18,7 @@ class GuideDialog extends StatelessWidget {
           if (child != null) child!,
           ActionBottomButton(
             onPressed: onPressed ?? () => Navigator.pop(context),
-            text: "확인",
+            text: t.confirm,
           )
         ],
       ),
