+++ lib/widgets/qrcode_link_info.dart
@@ -1,3 +1,4 @@
+import 'package:coconut_vault/localization/strings.g.dart';
 import 'package:flutter/material.dart';
 import 'package:coconut_vault/constants/app_info.dart';
 import 'package:qr_flutter/qr_flutter.dart';
@@ -77,19 +78,15 @@ class _QRCodeLinkInfoState extends State<QRCodeLinkInfo> {
                 textAlign: TextAlign.center),
             const SizedBox(height: 24),
             widget.qrData.contains('@')
-                ? widget.qrData.contains('라이선스')
+                ? widget.qrData.contains(t.license)
                     ? const Text(
                         'Please scan the QR code on a network-enabled device or send an email to the address above.\n\n네트워크가 활성화된 기기에서 QR 코드를 스캔하시거나 위의 주소로 메일을 전송해 주세요.',
                         style: Styles.body1,
                         textAlign: TextAlign.center)
-                    : const Text(
-                        '네트워크가 활성화된 기기에서 QR 코드를 스캔하시거나 위의 주소로 메일을 전송해 주세요.',
-                        style: Styles.body1,
-                        textAlign: TextAlign.center)
-                : const Text(
-                    '네트워크가 활성화된 기기에서 QR 코드를 스캔하시거나 위의 URL 주소로 접속해 주세요.',
-                    style: Styles.body1,
-                    textAlign: TextAlign.center),
+                    : Text(t.scan_qr_email_link,
+                        style: Styles.body1, textAlign: TextAlign.center)
+                : Text(t.scan_qr_url_link,
+                    style: Styles.body1, textAlign: TextAlign.center),
           ],
         ));
   }
